Fix a bug where trying to replace existing config files could cause an error.

+++ lib/util/files.py
@@ -15,7 +15,8 @@ def mkdir(path):
         sys.exit(1)
 
 def cp(source, dest):
-    shutil.copy(source, dest)
+    if not os.path.isfile(dest + '/' + os.path.basename(source)):
+        shutil.copy(source, dest)
 
 def read(source_file):
     with open(source_file) as f:
